Script update
- create script for measure tool
