+++ Geoportal-WebApp/app/teplosnabzhenie/main.js
@@ -2,31 +2,23 @@ import '../_src/styles/style.css';
 import '../_src/styles/mapstyle.css';
 import '../_src/styles/tablestyle.css';
 
-import {Feature, View} from 'ol';
+import {View} from 'ol';
 import {Map as OLMap} from 'ol';
 import {ScaleLine, defaults as defaultControls} from 'ol/control.js';
 import Overlay from 'ol/Overlay.js';
 
-// import loadLayers from '../_src/scripts/loadLayer';
+import {lineMeasureLayer, pointMeasureLayer, measuretool, measureClear} from '../_src/scripts/measureTools.js';
+
 import loadTileLayers from '../_src/scripts/loadTile';
 import initLayerControlGroup from '../_src/scripts/addControlGroup';
 import loadVectorLayers from '../_src/scripts/loadVectorLayer';
 
 import JSONDataHeat from './layers_heat.json'
 
-import VectorLayer from 'ol/layer/Vector';
-import VectorSource from 'ol/source/Vector';
-import { LineString, Point } from 'ol/geom';
-
-import {getLength} from 'ol/sphere';
-
 var map = null
 const container = document.getElementById('popup');
 const content = document.getElementById('popup-content');
 const closer = document.getElementById('popup-closer');
-
-const table = document.getElementById('measureTable');
-const resultElem = document.getElementById('measureRes');
 
 let mode = 2
 // 1 - получение инфы об объектах на карте
@@ -76,63 +68,6 @@ function showInfo(event) {
   }
 }
 
-var prevPoint = null
-var PointVectorSource = new VectorSource()
-var pointMeasureLayer = new VectorLayer({
-  source: PointVectorSource
-})
-var linevectorSource = new VectorSource()
-var lineMeasureLayer = new VectorLayer({
-  source: linevectorSource
-})
-var MeasureRes = 0
-
-function measuretool(event) {
-
-  let len = ""
-
-  var thing = new Point(event.coordinate)
-
-  if (prevPoint !== null)
-  {
-    var LineThing = new LineString([])
-    LineThing.appendCoordinate(prevPoint)
-    LineThing.appendCoordinate(event.coordinate)
-    var linefeature = new Feature({
-      name: "Thing",
-      geometry: LineThing
-    })
-
-    linevectorSource.addFeature( linefeature );
-    
-    len = getLength(LineThing)
-    MeasureRes += len
-    len = len.toFixed(3)
-  }
-
-  var feature = new Feature({
-    name: "Thing",
-    geometry: thing
-  })
-
-  PointVectorSource.addFeature( feature );
-
-  prevPoint = event.coordinate
-
-  table.innerHTML += "<div>" + event.coordinate[0].toFixed(3) + "</div>" + 
-  "<div>"+ event.coordinate[1].toFixed(3) +"</div>" + 
-  "<div>"+ len +"</div>";
-
-  resultElem.innerHTML = MeasureRes.toFixed(3) + " м";
-}
-function measureClear() {
-  table.innerHTML = ""
-  resultElem.innerHTML = ""
-  PointVectorSource.clear()
-  linevectorSource.clear()
-  MeasureRes = 0
-  prevPoint = null
-}
 
 function initMap() {
   map = new OLMap({
@@ -165,7 +100,6 @@ function initMap() {
   map.addLayer(lineMeasureLayer)
   map.addLayer(pointMeasureLayer)
 }
-
 initMap();
 
 
@@ -202,16 +136,18 @@ document.getElementById("menuButton").addEventListener("click", MapNavigationVCh
 // скрыть показать инструменты измерений
 function MeasureControlBoxVChange() {
   var x = document.getElementById("measureButtonsBox");
+  var y = document.getElementById("measureTableBox");
   // console.log(x)
   if (x.style.display === "none") {
       x.style.display = "";
+      y.style.display = "";
+      closeOverlay()
       mode = 2
   } else {
       x.style.display = "none";
+      y.style.display = "none";
+      measureClear()
       mode = 1
   }
-  console.log(mode)
 }
 document.getElementById("measureControlButton").addEventListener("click", MeasureControlBoxVChange);
-
-document.getElementById("MeasureClear").addEventListener("click", measureClear);
